app/Code/pyproject.toml
[project]
name = "echo-analyze"
version = "0.1.0"
description = "MF-CAS Portfolio Analyzer"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "python-multipart>=0.0.26",
    "pydantic",
    "PyJWT[crypto]",
    "cryptography>=46.0.7",
    "casparser-isin>=2025.2.28",
    "click>=7.0,<9.0",
    "colorama>=0.4.6,<0.5",
    "pdfminer-six==20251230",
    "python-dateutil>=2.8.2,<3",
    "rich>=13.5.2,<14",
    "pygments>=2.20.0",
    "openpyxl",
    "xlrd>=2.0.1",
]

[project.scripts]
app = "app.main:app"
